improvement: add HOC to concentrate drag-and-drop code
